🔨 [skip actions] Create a base class (breaks LSP)
